Refactor Location types to use double for coordinates
Replaced decimal with double for all latitude/longitude properties in Location, EstacionTerrestre, GasStationModel, and related classes. Updated parsing, construction, and usage of Location objects throughout the codebase. Adjusted Bounds and Earth utility classes to use double constants. Refactored OSRM and Valhalla routing services to support the new double-based Location structure. Improved XML documentation and fixed minor naming issues.

+++ src/Libs/GasStationPrices/Models/Minetur/EstacionTerrestre.cs
@@ -37,10 +37,10 @@ public readonly record struct EstacionTerrestre
     [J("Localidad")] public required string Localidad { get; init; }
 
     [J("Latitud")] public required string Latitud { get; init; }
-    public decimal Lat => decimal.Parse(Latitud, Core.Constants.Globalization.NumberFormatInfoES);
+    public double Lat => double.Parse(Latitud, Core.Constants.Globalization.NumberFormatInfoES);
 
     [J("Longitud (WGS84)")] public required string Longitud { get; init; }
-    public decimal Lon => decimal.Parse(Longitud, Core.Constants.Globalization.NumberFormatInfoES);
+    public double Lon => double.Parse(Longitud, Core.Constants.Globalization.NumberFormatInfoES);
 
     public Travel.Models.Location LatLng => new(Lat, Lon);
