+++ src/Libs/Travel/Services/Geocoding/Impl/NominatimGeocodingServiceImpl.cs
@@ -21,7 +21,7 @@ internal class NominatimGeocodingServiceImpl(Settings.GeocodingServiceApi api, M
                 : [.. restResponse
                     .Where(p => !string.IsNullOrWhiteSpace(p.Display_name))
                     .Where(p => p.Lat != 0 && p.Lon != 0)
-                    .Select(p => new ViewModels.Place(p.Display_name!, new Models.Location((decimal)p.Lat, (decimal)p.Lon)))
+                    .Select(p => new ViewModels.Place(p.Display_name!, new Models.Location(p.Lat, p.Lon)))
                 ];
         }
         catch (Exception e) when (logger.LogAndHandle(e, "Unexpected error")) { }
